Add failing test for Net::LDAP#open with nested queries
The nested query should match "user2", but with Net::LDAP#open, we see:

  <"user3"> expected but was <"user2">.

And likewise with the outer search:

  <["user1", "user2"]> expected but was <["user1", "user3"]>.

What's happening is that the nested query is being sent after the server
has already sent more data to be read for the outer search, so when the
inner search reads, it gets that result. And when it comes time for the
outer search to handle the second results, it gets reads the next result
off the wire, the result of the inner search.

I was also able to get an infinite loop just by *always* performing an
inner search instead of only searching on the first result.