Skip failing searchSize integration test, add explanatory test
This disables a failing test that won't pass without changing the
behavior of Net::LDAP#search (and Net::LDAP::Connection#search).

This is due to sizeLimitExceeded being treated as a failure instead
of a deliberate termination of search results. Instead of returning
the partial results, we currently return none.

cc @jch @schaary