Add rubocop to project
This adds https://github.com/bbatsov/rubocop to the project with an auto
generated `.rubocop_todo.yml`. The violations can be addressed over time without
breaking the build.

Also added is the `rake ci` task that runs `test` and `rubocop` only on mri. The
`rubotest` task is the same is `rake ci` but will run on all rubies, the default
rake task just runs test as per usual.

Conflicts:
	net-ldap.gemspec