Remove deprecated ConnectionRefusedError
ConnectionError.new was returning a deprecated ConnectionRefusedError,
this pattern was introduced for backward compatibility at some point,
but there seems to be no other usage of `ConnectionRefusedError` needed in
the codebase.

If we need to keep the class around for external backward
compatibility, I would recommend that we introduce a private-ish way to
initialize `ConnectionRefusedError` without raising the deprecation
warning.

I'm happy to PR such a change, but it seems that we could bump the
version and get rid of this deprecation in the library.