#431, Add `ostruct` as a dependency to the gemspec
This commit adds `ostruct` as an explicit dependency
in the net-ldap gemspec.

With the release of Ruby 3.3.5 and later versions,
users of net-ldap may encounter warnings related
to the use of `ostruct` if it is not declared as
a dependency. By including `ostruct`, we aim to
enhance clarity regarding the gem's requirements
and prevent any runtime issues related to missing
dependencies.